+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## [5.10.1](https://github.com/ni/systemlink-grafana-plugins/compare/v5.10.0...v5.10.1) (2026-09-16)
+
+### Bug Fixes
+
+* **Work-Item:** Resolve WorkItemsQueryBuilder filter persistence issue and standardize filter value mappings ([#754](https://github.com/ni/systemlink-grafana-plugins/issues/754)) ([a876d5e](https://github.com/ni/systemlink-grafana-plugins/commit/a876d5efd4ce00abb6a0ed41635fb885f6586acc))
+
+## [5.10.0](https://github.com/ni/systemlink-grafana-plugins/compare/v5.9.3...v5.10.0) (2026-09-16)
+
+### Features
+
+* **work-items:** Add support for resource row flattening and lookup ([#745](https://github.com/ni/systemlink-grafana-plugins/issues/745)) ([ef84c32](https://github.com/ni/systemlink-grafana-plugins/commit/ef84c3297bd62eb5cf4e3a8e411d4c786a5f3ba4))
+
+## [5.9.3](https://github.com/ni/systemlink-grafana-plugins/compare/v5.9.2...v5.9.3) (2026-09-15)
+
+### Bug Fixes
+
+* **work-items:** transform duration filters from days/hours to seconds ([#752](https://github.com/ni/systemlink-grafana-plugins/issues/752)) ([039ff0a](https://github.com/ni/systemlink-grafana-plugins/commit/039ff0aa58f6db587787aae25c410531541b917a))
+
+## [5.9.2](https://github.com/ni/systemlink-grafana-plugins/compare/v5.9.1...v5.9.2) (2026-09-15)
+
+### Bug Fixes
+
+* **work-items:** align query builder fields with API structure ([#750](https://github.com/ni/systemlink-grafana-plugins/issues/750)) ([3e42154](https://github.com/ni/systemlink-grafana-plugins/commit/3e42154212d167ebbfaf36fd86095113bddd3721))
+
+## [5.9.1](https://github.com/ni/systemlink-grafana-plugins/compare/v5.9.0...v5.9.1) (2026-09-15)
+
+### Bug Fixes
+
+* **constants:** update COMBOBOX_WIDTH to improve layout ([#753](https://github.com/ni/systemlink-grafana-plugins/issues/753)) ([f182c6e](https://github.com/ni/systemlink-grafana-plugins/commit/f182c6efc8d1d84b95f5a2480821a9b4b57af3e1))
+
+## [5.9.0](https://github.com/ni/systemlink-grafana-plugins/compare/v5.8.0...v5.9.0) (2026-09-15)
+
+### Features
+
+* **work-items:** Support properties output with lookup-resolved fields  ([#743](https://github.com/ni/systemlink-grafana-plugins/issues/743)) ([265eb13](https://github.com/ni/systemlink-grafana-plugins/commit/265eb132b488d3ff4c05824b3e0f38b19eb1000d))
+
+## [5.8.0](https://github.com/ni/systemlink-grafana-plugins/compare/v5.7.0...v5.8.0) (2026-09-11)
+
+### Features
+
+* **core:** centralize query error handling and duration formatting ([#747](https://github.com/ni/systemlink-grafana-plugins/issues/747)) ([f1da583](https://github.com/ni/systemlink-grafana-plugins/commit/f1da583f45c536b48b6e00f5910eabdcaf1477e6))
+
 ## [5.7.0](https://github.com/ni/systemlink-grafana-plugins/compare/v5.6.1...v5.7.0) (2026-09-10)
 
 ### Features